+++ src/objects/impls/reactor.rs
@@ -46,16 +46,15 @@ extern "C" {
     pub fn my(this: &Reactor) -> bool;
 
     /// The [`Owner`] of this reactor that contains the owner's username, or
-    /// `None` if it's currently not under a player's
-    /// control.
+    /// `None` if it's currently not under a player's control.
     ///
     /// [Screeps documentation](https://docs-season.screeps.com/api/#Reactor.owner)
     #[wasm_bindgen(method, getter)]
     pub fn owner(this: &Reactor) -> Option<Owner>;
 }
 
-impl HasNativeId for Reactor {
-    fn native_id(&self) -> JsString {
+impl HasId for Reactor {
+    fn js_raw_id(&self) -> JsString {
         Self::id_internal(self)
     }
 }